Refactored socialurls
