+++ apps/admin-x-settings/src/components/settings/general/UserDetailModal.tsx
@@ -17,7 +17,7 @@ import {RoutingModalProps, useRouting} from '@tryghost/admin-x-framework/routing
 import {User, canAccessSettings, hasAdminAccess, isAdminUser, isAuthorOrContributor, isEditorUser, isOwnerUser, useDeleteUser, useEditUser, useMakeOwner} from '@tryghost/admin-x-framework/api/users';
 import {getImageUrl, useUploadImage} from '@tryghost/admin-x-framework/api/images';
 import {useGlobalData} from '../../providers/GlobalDataProvider';
-import {validateBlueskyUrl, validateFacebookUrl, validateThreadsUrl, validateTwitterUrl} from '../../../utils/socialUrls';
+import {validateBlueskyUrl, validateFacebookUrl, validateThreadsUrl, validateTwitterUrl} from '../../../utils/socialUrls/index';
 
 const validators: Record<string, (u: Partial<User>) => string> = {
     name: ({name}) => {
